fix(money): Phase 2 money audit remediation — all critical and convention issues
Critical fixes:
- payroll-service: implement PayrollGrpcService (was missing — compliance audit
  returned AUDIT_SKIPPED on every call; KRA anomaly detection now functional)
- payroll-service: make PaySlip.helb NOT NULL (entity + V3 migration); closes
  NPE in PayrollRun.sumField() for any PaySlip with null helb
- payroll-service: V4 migration adds NOT NULL to all payroll_runs total columns
- analytics-service: null-guard BigDecimal params in PayrollSummary.create()
- frontend: basicSalary: number → string on Employee interface (IEEE 754 risk)

Convention fixes:
- compliance-service: extract PayrollGrpcClient infrastructure wrapper; refactor
  ComplianceAuditService to use constructor injection (removes @GrpcClient field)
- compliance-service: replace String.format("%.2f", BigDecimal) with toPlainString()
  to eliminate implicit doubleValue() conversion in audit log messages
- shared/andikisha-common: remove 6 dead methods from Money (percentage, max,
  isZero, isGreaterThan, isLessThan, kes(long)) — zero callers anywhere
- shared/andikisha-common: add MoneyTest (17 tests — was zero coverage on Money)

+++ services/tenant-service/src/test/java/com/andikisha/tenant/unit/PlanServiceTest.java
@@ -30,9 +30,9 @@ class PlanServiceTest {
 
     @Test
     void getAvailablePlans_returnsOnlyActivePlans() {
-        Plan starter = Plan.create("Starter", PlanTier.STARTER, Money.kes(2500),
+        Plan starter = Plan.create("Starter", PlanTier.STARTER, Money.kes(BigDecimal.valueOf(2500)),
                 25, 2, true, true, false, false, false, false, false);
-        Plan enterprise = Plan.create("Enterprise", PlanTier.ENTERPRISE, Money.kes(15000),
+        Plan enterprise = Plan.create("Enterprise", PlanTier.ENTERPRISE, Money.kes(BigDecimal.valueOf(15000)),
                 500, 20, true, true, true, true, true, true, true);
 
         when(planRepository.findByTenantIdAndActiveTrue("SYSTEM"))
